Fixed: from int to long long

+++ GeneralAlgebraTeamLabProject/TestFactorization/tst_testfactorization.cpp
@@ -30,25 +30,25 @@ TestFactorization::~TestFactorization()
 
 void TestFactorization::test_case1()
 {
-    std::vector<int> pn1 = PollardFactorization::factorize(91);
-    std::vector<int> ans1{7, 13};
+    std::vector<long long> pn1 = PollardFactorization::factorize(91);
+    std::vector<long long> ans1{7, 13};
     QCOMPARE(ans1, pn1);
 
-    std::vector<int> pn2 = PollardFactorization::factorize(4642135);
-    std::vector<int> ans2{5, 928427};
+    std::vector<long long> pn2 = PollardFactorization::factorize(4642135);
+    std::vector<long long> ans2{5, 928427};
     QCOMPARE(ans2, pn2);
 
-    std::vector<int> pn3 = PollardFactorization::factorize(4641);
-    std::vector<int> ans3{3, 7, 13, 17};
+    std::vector<long long> pn3 = PollardFactorization::factorize(4641);
+    std::vector<long long> ans3{3, 7, 13, 17};
     QCOMPARE(ans3, pn3);
 }
 
 void TestFactorization::test_case2()
 {
-    for (int i = 2; i <= 100000; ++i) {
+    for (long long i = 2; i <= 100000; ++i) {
         bool isExceptionThrown = false;
         try {
-            std::vector<int> pn = PollardFactorization::factorize(i);
+            std::vector<long long> pn = PollardFactorization::factorize(i);
         } catch(std::exception e) {
             isExceptionThrown = true;
         }
